requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
mangum>=0.17.0  # AWS Lambda handler for ASGI applications
zappa>=0.57.0   # AWS Lambda deployment tool

# Database
sqlalchemy>=1.4.23
alembic>=1.7.1
psycopg2-binary>=2.9.1

# AWS
boto3>=1.18.0
aws-lambda-powertools>=1.31.0

# ML and Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
# scikit-surprise>=1.1.3  # Temporarily commented out
python-dateutil>=2.8.2

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Development
python-dotenv>=0.19.0
black>=21.7b0
flake8>=3.9.2
isort>=5.9.3

# Jupyter
jupyter>=1.0.0
notebook>=6.4.3

# API Documentation
pydantic>=1.8.2 